#79: extracted StartScriptsMutator from RunTaskMutator
